refactor: extract `Packages` & `CompileFilter` to separate modules
